Wrapped each command in a function. Updated command calls in main to these functions
